+++ wire-ios-data-model/Source/Model/Conversation/ZMConversation+Message.swift
@@ -561,55 +561,3 @@ public extension ZMConversation {
     }
 
 }
-
-import UniformTypeIdentifiers
-
-public struct SendableImage {
-
-    public let name: String
-    public let utType: UTType?
-    public let data: Data
-
-    public init(
-        name: String?,
-        utType: UTType?,
-        data: Data
-    ) {
-        if let utType {
-            self.utType = utType
-        } else {
-            self.utType = Self.determineUTType(from: data)
-        }
-
-        if let name {
-            self.name = name
-        } else if let fileExtension = self.utType?.preferredFilenameExtension {
-            self.name = "picture.\(fileExtension)"
-        } else {
-            self.name = "picture"
-        }
-
-        self.data = data
-    }
-
-    private static func determineUTType(from data: Data) -> UTType? {
-        guard
-            !data.isEmpty,
-            let imageSource = CGImageSourceCreateWithData(data as CFData, nil),
-            let uti = CGImageSourceGetType(imageSource) as String?
-        else {
-            return nil
-        }
-
-        return UTType(uti)
-    }
-
-}
-
-extension SendableImage: CustomDebugStringConvertible {
-
-    public var debugDescription: String {
-        "SendableImage(name: \(name), type: \(utType?.preferredMIMEType ?? "unknown"))"
-    }
-
-}
